+minor fix
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@951 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/server_handlers.c
@@ -1076,7 +1076,7 @@ int _handle_create(struct rfsd_instance *instance, const struct sockaddr_in *cli
 	}
 	else
 	{
-		handle = htonll(fd);
+		handle = htonll((uint64_t)fd);
 		
 		if (rfs_send_answer_data(&instance->sendrecv, &ans, &handle) == -1)
 		{
